+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 master
 ------
 
+* Remove `bin/heroku_install` to simplify Heroku setup.
+  To upgrade, remove your project's generated `bin/heroku_install`.
+  Next, execute `rake ember:heroku`. [#544]
 * Generate an empty `yarn.lock` so that Heroku understands that the
   application's deployment target requires `yarn`. Closes [#538]. [#540]
 
+[#544]: https://github.com/thoughtbot/ember-cli-rails/pull/544
 [#538]: https://github.com/thoughtbot/ember-cli-rails/issues/538
 [#540]: https://github.com/thoughtbot/ember-cli-rails/pull/540